feat: 유저인지 체크하는 서비스 추가
테스트코드 수정

+++ src/test/java/kr/hhplus/be/server/user/UserTest.java
@@ -29,7 +29,7 @@ public class UserTest {
     @Test
     void isUserTest() {
         // given
-        final long userNo = 1;
+        final int userNo = 1;
 
         // when
         int count = userService.checkUserCountByUserNo(userNo);
